+++ nav-app/src/app/utils/globals.ts
