+++ nav-app/src/app/utils/navigator.d.ts
