+++ nav-app/src/app/utils/taxii2lib.ts
